Add FollowGameRepository to handle follow flow separately
